+++ src/folder.c
@@ -30,11 +30,11 @@ static int get_exec_path (char *exec_path, const size_t exec_path_sz)
 
   snprintf (tmp, sizeof (tmp) - 1, "/proc/%d/exe", getpid ());
 
-  const int len = readlink (tmp, exec_path, exec_path_sz - 1);
+  const ssize_t len = readlink (tmp, exec_path, exec_path_sz - 1);
 
   #elif defined (_WIN)
 
-  const int len = GetModuleFileName (NULL, exec_path, exec_path_sz - 1);
+  const DWORD len = GetModuleFileName (NULL, exec_path, exec_path_sz - 1);
 
   #elif defined (__APPLE__)
 
@@ -59,7 +59,7 @@ static int get_exec_path (char *exec_path, const size_t exec_path_sz)
 
   sysctl (mib, 4, exec_path, &size, NULL, 0);
 
-  const int len = readlink (tmp, exec_path, exec_path_sz - 1);
+  const ssize_t len = readlink (tmp, exec_path, exec_path_sz - 1);
 
   #else
   #error Your Operating System is not supported or detected
